Refactoring - 기타 폴더 정리, Study 필터 기능 오류 수정
