+++ pages/index.tsx
@@ -2,7 +2,7 @@ import { NextPage } from 'next'
 import MyInfoContainer from '@components/home/myInfoContainer'
 import Head from 'next/head'
 import ImageContainer from '@components/home/imageContainer'
-import UpButton from '@components/upButton'
+import UpButton from '@components/common/upButton'
 import MyStackContainer from '@components/home/myStackContainer'
 import MyProjectContainer from '@components/home/myProjectContainer'
