+++ pages/study/[id].tsx
@@ -10,7 +10,7 @@ import useMutation from '@libs/client/useMutation'
 import { useRouter } from 'next/router'
 import { useEffect, useRef } from 'react'
 import { Comment as CommentType, User } from '@prisma/client'
-import Button from '@components/button'
+import Button from '@components/common/button'
 import { contentType } from '@libs/client/notionContentType'
 import Head from 'next/head'
